backend/requirements.txt
# Core Framework
Django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
python-dotenv>=1.0.0
validators>=0.22.0

# Database & Queue
psycopg2-binary>=2.9.9
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
celery>=5.3.4

# Document Extraction (Pure Python)
pypdf>=3.17.0
python-docx>=1.1.0

# Web Scraping
beautifulsoup4>=4.12.2
requests>=2.31.0
readability-lxml>=0.8.1

# AI & LLM
google-generativeai>=0.3.2
tenacity>=8.2.0
pgvector>=0.2.0

# Utilities
gunicorn>=21.2.0
whitenoise>=6.6.0

# NOTE: Media processing libraries (rembg, whisper) removed
# to ensure compatibility with Python 3.14 on Windows.